save delay fixed

+++ app/src/main/java/com/example/passwordmanager/data/EncryptionHelper.kt
@@ -3,7 +3,6 @@ package com.example.passwordmanager.data
 import android.content.Context
 import android.security.keystore.KeyGenParameterSpec
 import android.security.keystore.KeyProperties
-import android.widget.Toast
 import androidx.security.crypto.EncryptedFile
 import androidx.security.crypto.MasterKey
 import com.google.gson.Gson
@@ -52,10 +51,7 @@ object EncryptionHelper {
 
     fun saveAccounts(context: Context, accounts: List<Account>) {
         val file = File(context.filesDir, FILE_NAME)
-        if (file.exists()) {
-            file.delete()
-            Toast.makeText(context, file.name, Toast.LENGTH_SHORT).show()
-        }
+        if (file.exists()) file.delete()
         val secretKey = getOrCreateSecretKey()
         val cipher = Cipher.getInstance(TRANSFORMATION)
         cipher.init(Cipher.ENCRYPT_MODE, secretKey)
